+++ HighPop/App.xaml.cs
@@ -196,6 +196,7 @@ public partial class App : System.Windows.Application
         s.AddSingleton<UserService>();
         s.AddSingleton<ServerGroupService>();
         s.AddSingleton<GroupBanListService>();
+        s.AddSingleton<RustModerationService>();
         s.AddSingleton<ServerHygieneService>();
         s.AddSingleton<WebApiService>();
         s.AddSingleton<RemoteMachineService>();